Refactor duplicates into its own module
